move validation methods to settings module; re-add kwargs
