+++ fmskill/plot.py
@@ -15,34 +15,17 @@ import fmskill.settings as settings
 from .settings import options, register_option
 
 
-def is_positive(value: object) -> None:
-    if np.isreal(value) and value > 0:
-        return
-    raise ValueError("Value must be a number greater than 0")
-
-
-def is_nonnegative(value: object) -> None:
-    if np.isreal(value) and value >= 0:
-        return
-    raise ValueError("Value must be a non-negative number")
-
-
-def is_between_0_and_1(value: object) -> None:
-    if np.isreal(value) and value >= 0 and value <= 1:
-        return
-    raise ValueError("Value must be a number between 0 and 1")
-
-
-register_option("plot.scatter.points.size", 20, validator=is_positive)
-register_option("plot.scatter.points.alpha", 0.5, validator=is_between_0_and_1)
+register_option("plot.scatter.points.size", 20, validator=settings.is_positive)
+register_option("plot.scatter.points.alpha", 0.5, validator=settings.is_between_0_and_1)
 register_option("plot.scatter.points.label", "", validator=settings.is_str)
 register_option("plot.scatter.quantiles.marker", "X", validator=settings.is_str)
-register_option("plot.scatter.quantiles.markersize", 3.5, validator=is_positive)
+register_option("plot.scatter.quantiles.markersize", 3.5, validator=settings.is_positive)
 register_option(
     "plot.scatter.quantiles.color", "darkturquoise", validator=settings.is_str
 )
 register_option("plot.scatter.quantiles.label", "Q-Q", validator=settings.is_str)
 register_option("plot.scatter.quantiles.markeredgecolor", (0, 0, 0, 0.4))
+register_option("plot.scatter.quantiles.kwargs", {}, settings.is_dict)
 register_option("plot.scatter.oneone_line.label", "1:1", validator=settings.is_str)
 register_option("plot.scatter.oneone_line.color", "blue", validator=settings.is_str)
 register_option("plot.scatter.reg_line.color", "r", validator=settings.is_str)
@@ -289,6 +272,7 @@ def scatter(
             zorder=4,
             markeredgecolor=options.plot.scatter.quantiles.markeredgecolor,
             markersize=options.plot.scatter.quantiles.markersize,
+            **settings.get_option("plot.scatter.quantiles.kwargs")
         )
         plt.plot(
             x,
